requirements.txt
Flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
python-dotenv==1.0.1
Markdown==3.5.2
pandas==2.2.1
joblib==1.3.2
scikit-learn==1.4.1.post1